% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_get_papers_info.R
\name{gld.get.papers.info}
\alias{gld.get.papers.info}
\title{Extracts information about papers from xml data}
\usage{
gld.get.papers.info(l.papers, name.author, id.author)
}
\arguments{
\item{l.papers}{A list with papers information}

\item{name.author}{Name of author}

\item{id.author}{ID of author}
}
\value{
A dataframe with several information about papers (published or accepted)
}
\description{
Extracts information about papers from xml data
}
\examples{

\dontrun{# INTERNAL USE
 tpublich <- gld.get.papers.info(i.papers, name.author, id.author)
}

}
