% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utils.R
\name{gld_download_lattes_files}
\alias{gld_download_lattes_files}
\title{Downloads data from Lattes}
\usage{
gld_download_lattes_files(id, folder.dl)
}
\arguments{
\item{id}{Id symbol from Lattes (see gld_get_lattes_data)}

\item{folder.dl}{Name of folder where to store xml files (default = tempdir())}
}
\value{
The name of downloaded zip file
}
\description{
Downloads data from Lattes
}
\examples{

\dontrun{
file.out <- gld_download_lattes_files(id = 'K4723925J2')
}

}
