% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utils.R
\name{gld_get_qualis}
\alias{gld_get_qualis}
\title{Reads Qualis table}
\usage{
gld_get_qualis(field.qualis = "ALL")
}
\arguments{
\item{field.qualis}{Area of Qualis to get Qualis journal rankings (default equals NULL). Eg. area.qualis <- 'ECONOMIA'}
}
\value{
A dataframe with Qualis table, including columns for ISSN, field and ranking
}
\description{
Read Qualis data available within the package. The original data is downloaded from
\url{https://sucupira.capes.gov.br/sucupira/public/index.xhtml}
}
\details{
Latest Qualis: 2013-2016 | Latest Qualis update: 2017-09-04
}
\examples{

df.qualis <- gld_get_qualis()
df.qualis
}
