% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utils.R
\name{gld_get_SJR}
\alias{gld_get_SJR}
\title{Reads SJR table}
\usage{
gld_get_SJR()
}
\value{
A dataframe with SJR table
}
\description{
Reads localy available SJR table. Original data provided as an excel file in \url{http://scimagojr.com/journalrank.php}.
}
\details{
Latest SJR: 2016 | Latest update: 2017-09-04
}
\examples{
df.SJR <- gld_get_SJR()
df.SJR
}
