% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_read_lattes_zip.R
\name{gld_read_zip}
\alias{gld_read_zip}
\title{Reads xml data from lattes zip file}
\usage{
gld_read_zip(zip.in)
}
\arguments{
\item{zip.in}{A (single) name of zip file containing a xml file}
}
\value{
A list with the following items:
 \describe{
  \item{tpesq}{A dataframe with information about researchers}
  \item{tpublic}{A dataframe with information about publications}
}
}
\description{
Reads xml data from lattes zip file
}
\examples{

f.in <- system.file('extdata/3262699324398819.zip', package = 'GetLattesData')
my.l <- gld_read_zip(f.in)
my.l
}
