% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_read_lattes_zip2.R
\name{gld_read_zip2}
\alias{gld_read_zip2}
\title{Reads xml data from lattes zip file}
\usage{
gld_read_zip2(f_zip)
}
\arguments{
\item{f_zip}{A (single) name of zip file containing a xml file}
}
\value{
A list with several items
}
\description{
Reads xml data from lattes zip file
}
\examples{

f_in <- system.file('extdata/3262699324398819.zip', package = 'GetLattesData')
my_l <- gld_read_zip2(f_in)
my_l
}
